style: view home,n -detail post

+++ src/helpers/date.js
@@ -1,7 +1,7 @@
 export function formatDate(value) {
   let date = new Date(value);
   const day = date.toLocaleString("default", { day: "2-digit" });
-  const month = date.toLocaleString("default", { month: "short" });
+  const month = date.toLocaleString("default", { month: "long" });
   const year = date.toLocaleString("default", { year: "numeric" });
   return `${month} ${day}, ${year}`;
 }
